# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ClawOPT backend",
    "version": "1.0.0",
    "description": "Generated from the backend route registry (npm run openapi:generate). Do not edit by hand."
  },
...
  "paths": {
...
    "/api/gateway/status": {
      "get": {
        "operationId": "getApiGatewayStatus",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/gateway/status",
        "x-clawopt-order": 39
      }
    },
    "/api/config/test": {
      "post": {
        "operationId": "postApiConfigTest",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/config/test",
        "x-clawopt-order": 40
      }
    },
    "/api/config/detect-all": {
      "get": {
        "operationId": "getApiConfigDetectAll",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/config/detect-all",
        "x-clawopt-order": 41
      }
    },
...
    "/api/config/browser-headed-mode": {
      "get": {
        "operationId": "getApiConfigBrowserHeadedMode",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/config/browser-headed-mode",
        "x-clawopt-order": 44
      },
      "post": {
        "operationId": "postApiConfigBrowserHeadedMode",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/config/browser-headed-mode",
        "x-clawopt-order": 47
      }
    },
    "/api/config/restart/status": {
      "get": {
        "operationId": "getApiConfigRestartStatus",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/config/restart/status",
        "x-clawopt-order": 45
      }
    },
    "/api/config/restart/status/reset": {
      "post": {
        "operationId": "postApiConfigRestartStatusReset",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/config/restart/status/reset",
        "x-clawopt-order": 46
      }
    },
...
    "/api/config/max-permissions": {
      "get": {
        "operationId": "getApiConfigMaxPermissions",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": false,
        "x-clawopt-express-path": "/api/config/max-permissions",
        "x-clawopt-order": 49
      },
      "post": {
        "operationId": "postApiConfigMaxPermissions",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/config/max-permissions",
        "x-clawopt-order": 50
      }
    },
    "/api/config/max-permissions/device-pairing/approve": {
      "post": {
        "operationId": "postApiConfigMaxPermissionsDevicePairingApprove",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/config/max-permissions/device-pairing/approve",
        "x-clawopt-order": 51
      }
    },
    "/api/config/restart": {
      "post": {
        "operationId": "postApiConfigRestart",
        "tags": [
          "control/gateway"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "authTokenHeader": []
          },
          {
            "bearerToken": []
          }
        ],
        "responses": {
          "default": {
            "description": "Response shape is not described yet (see the route handler)."
          }
        },
        "x-clawopt-public": false,
        "x-clawopt-admin-only": true,
        "x-clawopt-express-path": "/api/config/restart",
        "x-clawopt-order": 52
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "clawopt_session",
        "description": "httpOnly session cookie issued by POST /api/auth/login (web UI)."
      },
      "authTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ClawOPT-Auth-Token",
        "description": "Session token for CLI and scripts."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Same session token sent as Authorization: Bearer."
      }
    }
  }
}